tweaked styling, refactored Create User feature to use reusable component as HoC as well as fixing errors from createUser actions
